Move models to common package
